Split HttpUtils into CTH and API (Utils) versions
